+++ packages/marko/test/render/fixtures/await-client-reorder-with-runtimeId/test.js
@@ -1,0 +1,31 @@
+var expect = require("chai").expect;
+var extend = require("raptor-util/extend");
+
+exports.templateData = {
+  outer: Promise.resolve(),
+  inner1: Promise.resolve(),
+  inner2: Promise.resolve(),
+  $global: {
+    runtimeId: "testRuntimeId",
+    componentIdPrefix: "testComponentIdPrefix",
+  },
+};
+
+exports.checkEvents = function (events, snapshot, out) {
+  events = events.map(function (eventInfo) {
+    var arg = extend({}, eventInfo.arg);
+    expect(arg.out != null).to.equal(true);
+
+    delete arg.out; // Not serializable
+    delete arg.asyncValue; // Not serializable
+
+    return {
+      event: eventInfo.event,
+      arg: arg,
+    };
+  });
+
+  snapshot(events, out.isVDOM ? "-events-vdom.json" : "-events.json");
+};
+
+exports.skip_vdom = "client-reorder/placeholders are not supported in vdom";