+++ packages/translator-default/test/fixtures/at-tags/snapshots/html-expected.js
@@ -2,20 +2,19 @@ import { t as _t } from "marko/src/runtime/html/index.js";
 const _marko_componentType = "packages/translator-default/test/fixtures/at-tags/template.marko",
   _marko_template = _t(_marko_componentType);
 export default _marko_template;
-import _marko_self_iterator from "marko/src/runtime/helpers/self-iterator.js";
+import { a as _marko_repeatable_attr_tag, i as _marko_render_input } from "marko/src/runtime/helpers/attr-tag.js";
 import _hello from "./components/hello/index.marko";
 import _marko_tag from "marko/src/runtime/helpers/render-tag.js";
 import _marko_renderer from "marko/src/runtime/components/renderer.js";
 const _marko_component = {};
 _marko_template._ = _marko_renderer(function (input, out, _componentDef, _component, state, $global) {
-  _marko_tag(_hello, {
-    "foo": {
+  _marko_tag(_hello, _marko_render_input(() => {
+    _marko_repeatable_attr_tag("foo", {
       "renderBody": out => {
         out.w("Foo!");
-      },
-      [Symbol.iterator]: _marko_self_iterator
-    }
-  }, out, _componentDef, "0");
+      }
+    });
+  }), out, _componentDef, "0");
 }, {
   t: _marko_componentType,
   i: true,