+++ packages/translator-default/test/fixtures/attr-class/snapshots/vdomProduction-expected.js
@@ -3,16 +3,16 @@ const _marko_componentType = "TKoJdMQb",
   _marko_template = _t(_marko_componentType);
 export default _marko_template;
 import _marko_class_merge from "marko/dist/runtime/helpers/class-value.js";
-import _marko_createElement from "marko/dist/runtime/vdom/helpers/v-element.js";
-const _marko_node = _marko_createElement("div", {
+import _marko_constElement from "marko/dist/runtime/vdom/helpers/const-element.js";
+const _marko_node = _marko_constElement("div", {
   "class": "a b"
-}, "1", null, 0, 1);
-const _marko_node2 = _marko_createElement("div", {
+}, 0);
+const _marko_node2 = _marko_constElement("div", {
   "class": "a b c"
-}, "2", null, 0, 1);
+}, 0);
 import _customTag from "./components/custom-tag.marko";
 import _marko_tag from "marko/dist/runtime/helpers/render-tag.js";
-import _marko_self_iterator from "marko/dist/runtime/helpers/self-iterator.js";
+import { a as _marko_repeatable_attr_tag, i as _marko_render_input } from "marko/dist/runtime/helpers/attr-tag.js";
 import _marko_dynamic_tag from "marko/dist/runtime/helpers/dynamic-tag.js";
 import _marko_renderer from "marko/dist/runtime/components/renderer.js";
 import { r as _marko_registerComponent } from "marko/dist/runtime/components/registry.js";
@@ -36,21 +36,21 @@ _marko_template._ = _marko_renderer(function (input, out, _componentDef, _compon
   _marko_tag(_customTag, {
     "class": ["a", false, "b"]
   }, out, _componentDef, "4");
-  _marko_dynamic_tag(out, input.test, () => ({
-    "class": ["a", {
-      b: c,
-      d
-    }],
-    "test": {
+  _marko_dynamic_tag(out, input.test, () => _marko_render_input(() => {
+    _marko_repeatable_attr_tag("test", {
       "class": ["a", {
         b: c,
         d
       }],
       "renderBody": out => {
         out.t("Hello", _component);
-      },
-      [Symbol.iterator]: _marko_self_iterator
-    }
+      }
+    });
+  }, {
+    "class": ["a", {
+      b: c,
+      d
+    }]
   }), null, null, null, _componentDef, "5");
 }, {
   t: _marko_componentType,