+++ packages/translator-default/test/fixtures/doctype/snapshots/vdomProduction-expected.js
@@ -2,8 +2,8 @@ import { t as _t } from "marko/dist/runtime/vdom/index.js";
 const _marko_componentType = "VJrYycFN",
   _marko_template = _t(_marko_componentType);
 export default _marko_template;
-import _marko_createElement from "marko/dist/runtime/vdom/helpers/v-element.js";
-const _marko_node = _marko_createElement("head", null, "1", null, 1, 0).e("title", null, "2", null, 1, 0).t("Title of the document");
+import _marko_constElement from "marko/dist/runtime/vdom/helpers/const-element.js";
+const _marko_node = _marko_constElement("head", null, 1).e("title", null, 1).t("Title of the document");
 import _marko_renderer from "marko/dist/runtime/components/renderer.js";
 import { r as _marko_registerComponent } from "marko/dist/runtime/components/registry.js";
 _marko_registerComponent(_marko_componentType, () => _marko_template);