+++ packages/translator-default/test/fixtures/no-update-modifier-multiple/snapshots/htmlProduction-expected.js
@@ -8,9 +8,14 @@ import _marko_renderer from "marko/dist/runtime/components/renderer.js";
 const _marko_component = {};
 _marko_template._ = _marko_renderer(function (input, out, _componentDef, _component, state, $global) {
   out.w(`<div><input${_marko_props(out, _componentDef, {
-    pa: ["value"]
+    pa: {
+      value: 1
+    }
   })}${_marko_attr("value", input.defaultValue)}><input${_marko_props(out, _componentDef, {
-    pa: ["value", "checked"]
+    pa: {
+      value: 1,
+      checked: 1
+    }
   })} type=checkbox${_marko_attr("value", input.defaultValue)}${_marko_attr("checked", input.checked)}></div>`);
 }, {
   t: _marko_componentType,